requirements.txt
Django==1.5.1
South==0.8.1
psycopg2==2.5
Pillow==2.0.0
